fillOut

+++ pass.php
@@ -35,7 +35,7 @@ if ($valid && isset($_POST['passwort_vergessen'], $_POST['email']) && (int)$_POS
         'kKunde, cSperre'
     );
     if ($validRecaptcha === false) {
-        $alertHelper->addAlert(Alert::TYPE_ERROR, Shop::Lang()->get('accountLocked'), 'accountLocked');
+        $alertHelper->addAlert(Alert::TYPE_ERROR, Shop::Lang()->get('fillOut'), 'accountLocked');
     } elseif (isset($kunde->kKunde) && $kunde->kKunde > 0 && $kunde->cSperre !== 'Y') {
         $step     = 'passwort versenden';
         $customer = new Customer((int)$kunde->kKunde);
@@ -50,28 +50,19 @@ if ($valid && isset($_POST['passwort_vergessen'], $_POST['email']) && (int)$_POS
 } elseif ($valid && isset($_POST['pw_new'], $_POST['pw_new_confirm'], $_POST['fpwh'])) {
     if ($_POST['pw_new'] === $_POST['pw_new_confirm']) {
         $resetItem = Shop::Container()->getDB()->select('tpasswordreset', 'cKey', $_POST['fpwh']);
-        if ($resetItem) {
-            $dateExpires = new DateTime($resetItem->dExpires);
-            if ($dateExpires >= new DateTime()) {
-                $customer = new Customer((int)$resetItem->kKunde);
-                if ($customer->kKunde > 0 && $customer->cSperre !== 'Y') {
-                    $customer->updatePassword($_POST['pw_new']);
-                    Shop::Container()->getDB()->delete('tpasswordreset', 'kKunde', $customer->kKunde);
-                    header('Location: ' . $linkHelper->getStaticRoute('jtl.php') . '?updated_pw=true');
-                    exit();
-                }
-                $alertHelper->addAlert(
-                    Alert::TYPE_ERROR,
-                    Shop::Lang()->get('invalidCustomer', 'account data'),
-                    'invalidCustomer'
-                );
-            } else {
-                $alertHelper->addAlert(
-                    Alert::TYPE_ERROR,
-                    Shop::Lang()->get('invalidHash', 'account data'),
-                    'invalidHash'
-                );
+        if ($resetItem !== null && ($dateExpires = new DateTime($resetItem->dExpires)) >= new DateTime()) {
+            $customer = new Customer((int)$resetItem->kKunde);
+            if ($customer->kKunde > 0 && $customer->cSperre !== 'Y') {
+                $customer->updatePassword($_POST['pw_new']);
+                Shop::Container()->getDB()->delete('tpasswordreset', 'kKunde', $customer->kKunde);
+                header('Location: ' . $linkHelper->getStaticRoute('jtl.php') . '?updated_pw=true');
+                exit();
             }
+            $alertHelper->addAlert(
+                Alert::TYPE_ERROR,
+                Shop::Lang()->get('invalidCustomer', 'account data'),
+                'invalidCustomer'
+            );
         } else {
             $alertHelper->addAlert(
                 Alert::TYPE_ERROR,
